2024 day 18 part 1

+++ src/main/java/com/jeffrpowell/adventofcode/Point2DUtils.java
@@ -15,8 +15,8 @@ public class Point2DUtils
 {
 	private Point2DUtils() {}
     
-	public static double getManhattenDistance(Point2D pt1, Point2D pt2) {
-		return Math.abs(pt1.getY() - pt2.getY()) + Math.abs(pt1.getX() - pt2.getX());
+	public static long getManhattenDistance(Point2D pt1, Point2D pt2) {
+		return Double.valueOf(Math.abs(pt1.getY() - pt2.getY()) + Math.abs(pt1.getX() - pt2.getX())).longValue();
 	}
 	
 	public static double getEuclideanDistance(Point2D pt1, Point2D pt2) {
